+++ test/modules/test-driver/index.js
@@ -1,0 +1,4 @@
+// test driver
+module.exports.dummy = function () {
+
+};